- Add a confirmation alert when deleting a product - Add a success alert when the product has been deleted.

+++ crm-redux/src/components/Product.jsx
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 //Redux
 import { useDispatch } from "react-redux";
@@ -12,9 +13,21 @@ const Product = ({ product }) => {
   //  Confirmar si desea eliminarlo
   const confirmDeleteProduct = (id) => {
     //Preguntar al usuario
+    Swal.fire({
+      title: "Are you sure?",
+      text: "A product that is deleted cannot be recovered",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        //Pasarlo al action
+        dispatch(deleteProductAction(id));
+        }
+    });
 
-    //Pasarlo al action
-    dispatch(deleteProductAction(id))
   };
 
   return (
